add quiz-time-limit options

+++ src/main/java/kh/edu/cstad/stackquizapi/domain/Avatar.java
@@ -5,7 +5,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -23,7 +22,7 @@ public class Avatar {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Byte avatarNo;
+    private Integer avatarNo;
 
     @Column(name = "avatar_name", nullable = false)
     private String name;
